requirements.txt
asgiref==3.7.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap4==2022.1
cssbeautifier==1.14.9
Django==4.2.4
django-crispy-forms==2.0
django-phone-auth==0.3.1
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djlint==1.32.1
drf-yasg==1.21.7
EditorConfig==0.12.3
html-tag-names==0.1.2
html-void-elements==0.1.0
inflection==0.5.1
isort==5.12.0
jsbeautifier==1.14.9
json5==0.9.14
packaging==23.1
pathspec==0.11.2
phonenumbers==8.13.18
psycopg2==2.9.7
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
